cast

+++ src/filesystem.rs
@@ -24,11 +24,11 @@ pub trait FileSystem {
 
 bitflags! {
     struct FileSystemFlags: types::c_int {
-        const FS_REQUIRES_DEV = bindings::FS_REQUIRES_DEV;
-        const FS_BINARY_MOUNTDATA = bindings::FS_BINARY_MOUNTDATA;
-        const FS_HAS_SUBTYPE = bindings::FS_HAS_SUBTYPE;
-        const FS_USERNS_MOUNT = bindings::FS_USERNS_MOUNT;
-        const FS_RENAME_DOES_D_MOVE = bindings::FS_RENAME_DOES_D_MOVE;
+        const FS_REQUIRES_DEV = bindings::FS_REQUIRES_DEV as types::c_int;
+        const FS_BINARY_MOUNTDATA = bindings::FS_BINARY_MOUNTDATA as types::c_int;
+        const FS_HAS_SUBTYPE = bindings::FS_HAS_SUBTYPE as types::c_int;
+        const FS_USERNS_MOUNT = bindings::FS_USERNS_MOUNT as types::c_int;
+        const FS_RENAME_DOES_D_MOVE = bindings::FS_RENAME_DOES_D_MOVE as types::c_int;
     }
 }
